% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundFun.R
\name{noisyMCMC}
\alias{noisyMCMC}
\title{Create sounds from MCMC outputs}
\usage{
noisyMCMC(my.mcmc, mean.freq = 400, speed = 0.01)
}
\arguments{
\item{my.mcmc}{dataframe containing all mcmc outputs. Each column should be a parameter, and the rows should be their estimated values.}

\item{mean.freq}{The mean frequency to use. Values of the MCMC adjust this mean frequency.}

\item{speed}{The speed at which sounds are played.}

\item{player}{The name of the player to be used by tuneR (Default: mplay32.exe). See tuneR ?play() for options.}
}
\description{
This function will generate sounds using the outputs of an MCMC.
}
\examples{
\dontrun{
#example MCMC outputs from a 3 parameter model
good.mcmc <- data.frame(x1 = rnorm(300,0,1),
                        x2 = rnorm(300,10,1),
                        x3 = rnorm(300,20,1) )
bad.mcmc <- data.frame(x1 = rnorm(300,0,1),
                       x2 = rnorm(300,10,1), 
                       x3 = c(rnorm(25,0,1),rnorm(25,10,0.1),rnorm(250,0,10)) )

#lisen to MCMC chain
noisyMCMC(good.mcmc)
noisyMCMC(bad.mcmc)
}     
}
